Added in file manifest handling and upgraded version

+++ src/configure.js
@@ -60,20 +60,11 @@ parser.addArgument(
   }
 );
 parser.addArgument(
-  ['-s', '--serve'],
-  {
-    help: 'Serve your documentation after setup',
-    action: 'storeTrue',
-    defaultValue: false,
-    dest: 'serveDocs'
-  }
-);
-parser.addArgument(
   ['--no-code'],
   {
     help: 'If true will not set up code docs with jsdoc',
-    action: 'storeTrue',
-    defaultValue: false,
+    action: 'storeFalse',
+    defaultValue: true,
     dest: 'codeDocs'
   }
 );
@@ -81,8 +72,8 @@ parser.addArgument(
   ['--no-react'],
   {
     help: 'If true will not set up react component style guide',
-    action: 'storeTrue',
-    defaultValue: false,
+    action: 'storeFalse',
+    defaultValue: true,
     dest: 'reactDocs'
   }
 );
@@ -148,7 +139,6 @@ async function printHeadingAndArt() {
   });
 };
 
-
 /**
  * Add script commands to the passed package json object.
  *
@@ -156,7 +146,16 @@ async function printHeadingAndArt() {
  * @param  {String} path Path to write updated package.json out to
  * @return {Promise}     Promise, when resolved has updated package.json
  */
-async function updatePackageJson(pkg, path) {
+
+/**
+ * Add script commands to the passed package json object.
+ * @param  {Object} pkg  { package.json }
+ * @param  {String} path Path to write updated package.json out to
+ * @param  {Boolean} codeDocs  True if include code docs scripts
+ * @param  {Boolean} reactDocs True if include react style guide scripts
+ * @return {Promise}     Promise, when resolved has updated package.json
+ */
+async function updatePackageJson(pkg, path, codeDocs, reactDocs) {
   return new Promise(function (resolve, reject) {
     try {
       if (!pkg.scripts) {
@@ -165,14 +164,27 @@ async function updatePackageJson(pkg, path) {
       console.log('');
       console.log(`${chalk.yellow('Adding docs and lint scripts to package.json.')}`);
 
-      pkg.scripts['docs'] = 'run-s -s docs:code docs:proj';
-      pkg.scripts['docs:code'] = 'run-s docs:clear docs:build docs:clear';
-      pkg.scripts['docs:proj'] = 'docsify serve docs';
-      pkg.scripts['docs:clear'] = 'rm -rf ./out/';
-      pkg.scripts['docs:build'] = 'jsdoc ./src || exit 0 && skan-convert -p ./out -o ./docs || exit 0';
-      pkg.scripts['lint'] = 'run-s -s lint:*';
-      pkg.scripts['lint:md'] = 'remark -i .gitignore --no-stdout --use remark-lint *.md';
-      pkg.scripts['lint:docs'] = 'remark -i .gitignore /docs --no-stdout --use remark-lint docs/*.md';
+      if (codeDocs && reactDocs) {
+        pkg.scripts['docs'] = 'run-s -s docs:code docs:proj';
+        pkg.scripts['docs:code'] = 'run-s docs:clear docs:build docs:clear';
+        pkg.scripts['docs:proj'] = 'docsify serve docs';
+        pkg.scripts['docs:clear'] = 'rm -rf ./out/';
+        pkg.scripts['docs:build'] = 'jsdoc ./src || exit 0 && skan-convert -p ./out -o ./docs || exit 0';
+        pkg.scripts['lint'] = 'run-s -s lint:*';
+        pkg.scripts['lint:md'] = 'remark -i .gitignore --no-stdout --use remark-lint *.md';
+        pkg.scripts['lint:docs'] = 'remark -i .gitignore /docs --no-stdout --use remark-lint docs/*.md';
+
+        // TODO add react style guide scripts
+
+      } else if (!codeDocs) {
+        pkg.scripts['docs'] = 'run-s -s docs:proj';
+        pkg.scripts['docs:proj'] = 'docsify serve docs';
+        pkg.scripts['lint'] = 'run-s -s lint:*';
+        pkg.scripts['lint:md'] = 'remark -i .gitignore --no-stdout --use remark-lint *.md';
+        pkg.scripts['lint:docs'] = 'remark -i .gitignore /docs --no-stdout --use remark-lint docs/*.md';
+
+        // TODO add react style guide scripts
+      }
 
       fs.writeFile(path, JSON.stringify(pkg, null, 2), (err)=> {
         if (err) {
@@ -219,7 +231,7 @@ async function copyFiles(packagePath) {
 async function main() {
   const args = parser.parseArgs(process.argv.slice(2));
 
-  const {packagePath, serveDocs, codeDocs, reactDocs} = args;
+  const {packagePath, codeDocs, reactDocs} = args;
 
   const pkgPath = packagePath === null
     ? path.resolve('.', 'package.json')
@@ -243,7 +255,7 @@ async function main() {
   console.log('');
   console.log(`${chalk.yellow('==> Updating package.json scripts...')}`);
 
-  await updatePackageJson(pkgJson, pkgPath);
+  await updatePackageJson(pkgJson, pkgPath, codeDocs, reactDocs);
 
   console.log('');
   console.log(`${chalk.yellow('==> Downloading and installing docsify-cli...')}`);
@@ -251,18 +263,22 @@ async function main() {
   await tryLogCommand('npm i -g docsify-cli && npm i -D docsify-cli');
 
   console.log(`${chalk.yellow('docsify-cli successfully installed.')}`);
-  console.log('');
-  console.log(`${chalk.yellow('==> Downloading and installing babel-core and babel-cli')}`);
 
-  await tryLogCommand('npm i babel-core babel-cli');
+  if (codeDocs) {
+    console.log('');
+    console.log(`${chalk.yellow('==> Downloading and installing babel-core and babel-cli')}`);
 
-  console.log(`${chalk.yellow('babel successfully installed.')}`);
-  console.log('');
-  console.log(`${chalk.yellow('==> Downloading and installing jsdoc and jsdoc-babel')}`);
+    await tryLogCommand('npm i babel-core babel-cli');
 
-  await tryLogCommand('npm i -g jsdoc jsdoc-babel && npm i -D jsdoc jsdoc-babel');
+    console.log(`${chalk.yellow('babel successfully installed.')}`);
+    console.log('');
+    console.log(`${chalk.yellow('==> Downloading and installing jsdoc and jsdoc-babel')}`);
 
-  console.log(`${chalk.yellow('jsdoc successfully installed.')}`);
+    await tryLogCommand('npm i -g jsdoc jsdoc-babel && npm i -D jsdoc jsdoc-babel');
+
+    console.log(`${chalk.yellow('jsdoc successfully installed.')}`);
+  }
+
   console.log('');
   console.log(`${chalk.yellow('==> Downloading and installing remark-cli and remark-lint...')}`);
